backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyMySQL==1.1.0
python-dotenv==1.0.1
django-cors-headers==4.3.1
drf-spectacular==0.27.2
